Bug fix: type of binary connectivity matrix is float to enable adding weights if needed

+++ conn2res/reservoir.py
@@ -109,7 +109,7 @@ class Conn:
         """
 
         # binarize connectivity matrix
-        self.w = self.w.astype(bool).astype(int)
+        self.w = self.w.astype(bool).astype(float)
 
     def add_weight(self, w, mask='triu'):
         """
